fix: add missing break in HubService switch case
Additional changes:
- enable javac lint warnings (except processing) in all subprojects via build.gradle.kts
- add serialVersionUID to OrmException and ParamException for serialization compatibility
- remove redundant int cast in RtvService vote count calculation
- refactor Column.in/notIn varargs methods to use ArrayList instead of List.of
- modernize CompletableFuture.toArray calls in AsyncExecutionTest tests
- simplify ApiGateway message handling by extracting a helper method, using proper generics and eliminating unchecked warnings

+++ plugin/src/main/java/plugin/vote/RtvService.java
@@ -161,7 +161,7 @@ public class RtvService {
             return 0;
         }
 
-        return (int) vote.count(uuid -> {
+        return vote.count(uuid -> {
             var session = sessionService.getByUuid(uuid);
             return session.isEmpty() || !session.get().isAfk();
         });
